Extracting the rest of the enums/structs, migrating usages
