+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/TdsParser.Windows.cs
@@ -4,6 +4,7 @@
 
 using System;
 using System.Diagnostics;
+using Interop.Windows.Sni;
 using Microsoft.Data.SqlClient.SNI;
 
 namespace Microsoft.Data.SqlClient
@@ -74,7 +75,7 @@ namespace Microsoft.Data.SqlClient
             }
             else
             {
-                SNINativeMethodWrapper.SNIGetLastError(out SNINativeMethodWrapper.SNI_Error sniError);
+                SNINativeMethodWrapper.SNIGetLastError(out SniError sniError);
                 details.sniErrorNumber = sniError.sniError;
                 details.errorMessage = sniError.errorMessage;
                 details.nativeError = sniError.nativeError;
